requirements.txt
google-genai
PyPDF2
requests
python-dotenv
